Fix deadlock
There was a deadlock if the managed subsystem is in state recover_very_high and the sonar fails.
This is fixed now by including another transition that can be taken in this case. It will either stay in recover_very_high or go to search_high (even if we don't know which altitude feature is active). From there on, it will switch to the correct altitude in the next transition.